Use window.location.origin to create InstalledApp url
Previously, the code used http://${window.location.hostname} to build
the base URL for checking the status of an app. Now it uses
${window.location.origin}.

In my case, I have set up nginx as a reverse proxy in order to use TLS
encryption for my node. I am accessing Umbrel via https:// now. As a
result, the functionality to check the status of services was not
working due to the browser blocking mixed content requests.